+++ src/app/np-dialog-demo/np-dialog-demo.component.ts
@@ -12,24 +12,30 @@ import { NpModalModule } from 'np-ui-lib';`;
 
   serviceText = `constructor(private modalService: NpModalService) { }`;
 
-  componentText = `<span class="np-text-success">// For alert dialog, pass type="alert"</span>
+  componentText = `<span class="np-text-success">// ----For alert dialog, pass type="alert"</span>
 this.modalService.open(NpDialogComponent, null, { type: 'alert', message: 'Saved successfully.' });
 
-<span class="np-text-success">// For prompt dialog, pass type="prompt"</span>
+<span class="np-text-success">// ----For prompt dialog, pass type="prompt"</span>
 const prompt = this.modalService.open(NpDialogComponent, null, { type: 'prompt', message: 'Enter your name' });
 prompt.onClose.subscribe((data) => {
-  <span class="np-text-success">// returns value in data, if ok button is clicked else undefined</span>
-  alert('Prompt value is ' + data);
+  <span class="np-text-success">// if Ok button is clicked then returns {action : 'Ok', value: string }</span>
+  <span class="np-text-success">// if Cancel button is clicked then returns {action : 'Cancel' }</span>
+  if (data.action === 'Ok') {
+    alert('Ok button is clicked and input value is : ' + data.value);
+  } else {
+    alert('Cancel button is clicked');
+  }
 });
 
-<span class="np-text-success">// For confirm dialog, pass type="confirm"</span>
+<span class="np-text-success">// ----For confirm dialog, pass type="confirm"</span>
 const confirm = this.modalService.open(NpDialogComponent, null, { type: 'confirm', message: 'Are you sure to delete?' });
 confirm.onClose.subscribe((data) => {
-  <span class="np-text-success">// returns true in data, if ok button is clicked</span>
-  if (data === true) {
-    alert('Ok clicked');
+  <span class="np-text-success">// if Ok button is clicked then returns {action : 'Ok'}</span>
+  <span class="np-text-success">// if Cancel button is clicked then returns {action : 'Cancel' }</span>
+  if (data.action === 'Ok') {
+    alert('Ok button is clicked');
   } else {
-    alert('Cancel clicked');
+    alert('Cancel button is clicked');
   }
 });
     `;
@@ -48,8 +54,7 @@ confirm.onClose.subscribe((data) => {
     const prompt = this.modalService.open(NpDialogComponent,
       null, { type: 'prompt', message: 'Enter your name' });
     prompt.onClose.subscribe((data) => {
-      // returns value in data, if ok button is clicked else undefined
-      alert('Prompt value is ' + data);
+      alert(JSON.stringify(data));
     });
   }
 
@@ -57,12 +62,7 @@ confirm.onClose.subscribe((data) => {
     const confirm = this.modalService.open(NpDialogComponent,
       null, { type: 'confirm', message: 'Are you sure to delete?' });
     confirm.onClose.subscribe((data) => {
-      // returns true in data, if ok button is clicked
-      if (data === true) {
-        alert('Ok clicked');
-      } else {
-        alert('Cancel clicked');
-      }
+      alert(JSON.stringify(data));
     });
   }
 }